task 0

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -4,14 +4,14 @@
 
 /**
  * print_listint - prints all elements of a listint_t
- * @h: the pointer to next node
+ * @h: the head of the list to print
  *
  * Return: the number of nodes
  */
 
 size_t print_listint(const listint_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 
 	while (h)
 	{
